Add tpm2-abrmd policy
The tpm2-abrmd is a user-space D-Bus based TPM2 Access Broker and Resource
Manager. Applications communicate with the tpm2-abrmd by sending commands
and receiving responses using a communication library (TCTI) that used to
create a pipe(2) to send the data to the daemon.

But this communication mechanism was changed to use unix sockets instead
and pass the file descriptors to the D-Bus daemon, so a SELinux policy is
needed to update the tpm2-abrmd to the latest version that use sockets.

This patch is based on the SELinux policy module that's in the tpm2-abrmd
upstream repository, with the changes suggested by Dominick Grift.

Suggested-by: Dominick Grift <dac.override@gmail.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>